Removing initial Element Timing implicit registration heuristic.
The Element Timing specification originally included a provision for reporting
timings of elements that were automatically detected as 'important'. This would
allow useful inforation from the Element Timing API to be reported through the
Performance Timeline even if site authors neglected to mark their important
elements with the 'elementtiming' attribute.

The automatic detection of 'important' elements is, necessarily, a heuristic
and, until consensus for the correct heuristic had been reached, we were
treating any image that occupied greater than %15 of the viewport as
'automatically important'. As the Element Timing specification is still being
refined, the exact threshold and heuristic is still being investigated. Until
agreement on an appropriate heuristic is reached, implicit registration will be
disabled.

This CL removes our %15 threshold and disables implicit registration. There was
also WPT coverage to test that all browsers would respect the %15 threshold so
that test is being removed too.

Change-Id: I14847533402f5a68bb3de63a8f9acbd004c5a8e4