Prevent leaking Sec-CH-/Sec-Fetch- Request Headers on HTTPS Downgrade Redirects.
Currently various Sec-CH- and Sec-Fetch- request headers are added through the codebase and correctly check if the request target url is trustworthy. If the destination is not trustworthy then the headers are not added. However in the event a request redirects from a trustworthy to non-trustworthy destination, such as https => http, there may be pre-existing headers present. Since these headers would not have been added to the non-trustworthy request we need to remove these extra headers from the request.

This change adds a helper call to the UrlLoader class in the network service to remove any request headers prefixed with "sec-ch-" or "sec-fetch-" when a downgrade redirect is detected. In addition to this unit test cases for the helper and end to end WPT cases have been added to validate the scenario and prevent future regressions.

Bug: 964053
Change-Id: I109c3ec1b3a05f6341c3c4adbd1a8da1274fd0d6