CSS Filter: Default value for filter functions
https://drafts.fxtf.org/filter-effects/#funcdef-filter-brightness
https://drafts.fxtf.org/filter-effects/#funcdef-filter-contrast
https://drafts.fxtf.org/filter-effects/#funcdef-filter-invert
https://drafts.fxtf.org/filter-effects/#funcdef-filter-opacity
> Default value when omitted is 1.

Consolidate the default value code for filter computed value and animation.
We previously used the wrong default value for contrast and opacity only
when these occurred in animation keyframes.

Includes WPTs.

This CL has a change in behavior, but brings us into consistency
with spec, Firefox and Safari.
Usage of filter with default arguments is < 0.04%

Bug: 964696
Change-Id: I78c238570fc23252795e60076d9d3ffa1aa91e6e